mv funcs to z_distributions
