+++ clmm/z_distributions.py
@@ -1,4 +1,30 @@
-from .utils import _srd_z_distrib, _chang_z_distrib
+"""Redshift distribution functions"""
+
+def _functional_form(redshift, alpha, beta, redshift0, is_cdf=False):
+    """
+    A private function that returns the functionnal form of the redshift distribution used in Chang et al (2013):
+
+    .. math::
+       P(z) = z^{\\alpha}\times\exp^{\left(-\frac{z}{z0}^\beta\right)}
+
+    Parameters
+    ----------
+    redshift : float
+        Galaxy redshift
+    alpha, beta, z0 : floats
+        Parameters describing the function
+    is_cdf : bool
+        If True, returns cumulative function.
+
+    Returns
+    -------
+    The value of the function at z
+    """
+    if is_cdf:
+        return redshift0**(alpha+1)*gammainc((alpha+1)/beta, (redshift/redshift0)**beta)/beta*gamma((alpha+1)/beta)
+    else:
+        return (redshift**alpha)*np.exp(-(redshift/redshift0)**beta)
+
 
 def chang2013(redshift, is_cdf=False):
     """
@@ -16,7 +42,9 @@ def chang2013(redshift, is_cdf=False):
     -------
     The value of the distribution at z
     """
-    return _chang_z_distrib(redshift, is_cdf)
+    alpha, beta, redshift0 = 1.24, 1.01, 0.51
+    return _functional_form(redshift, alpha, beta, redshift0, is_cdf)
+
 
 def desc_srd(redshift, is_cdf=False):
     """
@@ -34,4 +62,5 @@ def desc_srd(redshift, is_cdf=False):
     -------
     The value of the distribution at z
     """
-    return _srd_z_distrib(redshift, is_cdf)
+    alpha, beta, redshift0 = 2., 0.9, 0.28
+    return _functional_form(redshift, alpha, beta, redshift0, is_cdf)
